% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{bootstrapReliability}
\alias{bootstrapReliability}
\title{Simulate the reliability}
\usage{
bootstrapReliability(decisions)
}
\arguments{
\item{decisions}{Decisions from nmmMongo}
}
\value{
reliability values
}
\description{
Simulate the reliability
}
