% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{btm_fit_2}
\alias{btm_fit_2}
\title{Estimate judge infit}
\usage{
btm_fit_2(probs, decisions)
}
\arguments{
\item{decisions}{Decisions from nmmMongo}

\item{probabilities}{Probabilities of decisions from btm model}
}
\value{
data frame of judges and infit values
}
\description{
Estimate judge infit
}
\examples{
mdl <- btmModel(decisions)
probs <- mdl$probs
judge.infit <- btm_fit_2(probs,decisions)
}
