% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{btmBootstrap}
\alias{btmBootstrap}
\title{Bootstrap}
\usage{
btmBootstrap(decisions, iterations = 10, size = 1, replace = TRUE)
}
\arguments{
\item{decisions}{decisions from nmmMongo}

\item{iterations}{number of boostrap iterations}

\item{size}{size of bootstrap sample as a fraction}

\item{replace}{boostrap with or without replacement}
}
\value{
Data frame of thetas
}
\description{
Bootstrap
}
