% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{btmModel}
\alias{btmModel}
\title{Fit the BTM Model to data from nmmMongo}
\usage{
btmModel(decisions, anchors = NULL)
}
\arguments{
\item{decisions}{Decisions from nmmMongo}

\item{anchors}{Named list of anchor names and values}
}
\value{
Bradley Terry model
}
\description{
Fit the BTM Model to data from nmmMongo
}
\examples{
mdl <- btmModel(decisions)
}
