% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{btmSimulate}
\alias{btmSimulate}
\title{Simulate}
\usage{
btmSimulate(decisions, iterations = 10)
}
\arguments{
\item{decisions}{decisions from nmmMongo}

\item{iterations}{number of simulations}
}
\value{
Data frame of thetas
}
\description{
Simulate
}
