% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{expP}
\alias{expP}
\title{Probability of outcome}
\usage{
expP(a, b)
}
\arguments{
\item{a}{ability of person a}

\item{b}{ability of person b}
}
\value{
probability of outcome
}
\description{
Probability of outcome
}
