% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{interrater}
\alias{interrater}
\title{Estimate inter-rater reliability using split-halves}
\usage{
interrater(iterations, decisions)
}
\arguments{
\item{iterations}{Number of iterations required (typically 100)}

\item{decisions}{Decisions from NMM}
}
\value{
Median inter-rater reliability
}
\description{
Estimate inter-rater reliability using split-halves
}
\examples{
r <- interrater(iterations,decisions)
}
