% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{passFail}
\alias{passFail}
\title{Pass / Fail classification accuracy}
\usage{
passFail(cutOff, persons)
}
\arguments{
\item{cutOff}{pass/fail theta}

\item{persons}{persons from nmmMongo}
}
\value{
classification accuracy list with summary values & all person values
}
\description{
Pass / Fail classification accuracy
}
