% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{relJudgeExcluded}
\alias{relJudgeExcluded}
\title{Reliability with judges excluded}
\usage{
relJudgeExcluded(decisions)
}
\arguments{
\item{decisions}{Decisions from nmmMongo}
}
\value{
data frame with judge & reliability if excluded
}
\description{
Reliability with judges excluded
}
