% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{scaleThetas}
\alias{scaleThetas}
\title{Scaled scores}
\usage{
scaleThetas(mdlEffects, wRange, wlow)
}
\arguments{
\item{mdlEffects}{Effects from btm model}

\item{wRange}{Wanted range of scaled scores}

\item{wlow}{Wanted low of scaled scores}
}
\value{
Data frame of scaled scores and standard errors
}
\description{
Scaled scores
}
\examples{
mdl <- btmModel(decisions)
scaledScores <- scaleThetas(modl$effects, 20, 0)
}
