% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{visReliability}
\alias{visReliability}
\title{Visualise the reliability}
\usage{
visReliability(decisions)
}
\arguments{
\item{decisions}{Decisions from nmmMongo}
}
\value{
reliability plot
}
\description{
Visualise the reliability
}
