% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{bootstrapReliability}
\alias{bootstrapReliability}
\title{Simulate the reliability if judged to 100%}
\usage{
bootstrapReliability(decisions, judges)
}
\arguments{
\item{decisions}{Decisions from nmmMongo}

\item{judges}{Judges from nmmMongo}
}
\value{
bootstrapped reliability when judged to 100%
}
\description{
Simulate the reliability if judged to 100%
}
