% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm-functions.R
\name{btmSimulate}
\alias{btmSimulate}
\title{Simulate}
\usage{
btmSimulate(decisions, anchors, iterations = 10)
}
\arguments{
\item{decisions}{decisions from nmmMongo}

\item{iterations}{number of simulations}

\item{anchor}{named list of anchor values}
}
\value{
Data frame of thetas
}
\description{
Simulate
}
\examples{
data("decisions")
}
